перенос компонента
